refactor: [IMP] update api types and queries
